pyproject.toml
[tool.poetry]
name = "ping-the-web"
version = "0.1.0"
description = ""
authors = ["Kloven <jloup.m@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
tqdm = "^4.64.0"
hilbertcurve = "^2.0.5"
Pillow = "^9.3.0"


[tool.poetry.dev-dependencies]
black = { version = "^22.3.0", allow-prereleases = true }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
